package support

+++ main.c
@@ -3,6 +3,8 @@
 #include "koala.h"
 #include "parser.h"
 #include "codegen.h"
+#include "koala_yacc.h"
+#include "koala_lex.h"
 
 #define KOALA_VERSION "0.7.1"
 
@@ -117,30 +119,80 @@ static int cmd_build(char *argv[], int size)
 
   char *input = argv[0];
 
-  FILE *in = fopen(input, "r");
-  if (!in) {
+  struct stat sb;
+  if (lstat(input, &sb) == - 1) {
+    printf("%s: invalid filename\n", input);
+    return 0;
+  }
+
+  if (!S_ISDIR(sb.st_mode)) {
+    printf("%s: is not a directory\n", input);
+    return 0;
+  }
+
+  DIR *dir = opendir(input);
+  if (!dir) {
     printf("%s: no such file or directory\n", input);
     return 0;
   }
 
-  int len = strlen(input) - 2;
-  char output[len + 4];
-  memcpy(output, input, len);
-  strcpy(output + len, "klc");
+  int len = strlen(input);
+  char output[16 + len + 4];
+  strcpy(output, "../pkg/");
+  strncpy(output + 7, input, len);
+  strcpy(output + len + 7, ".klc");
+  printf("path:%s\n", output);
 
-  ParserState *ps;
   Koala_Initialize();
-  ps = new_parser(input);
-  parser_module(ps, in);
-  if (ps->errnum <= 0) {
-    codegen_klc(ps, output);
-  } else {
-    fprintf(stderr, "There are %d errors.\n", ps->errnum);
-  }
-  destroy_parser(ps);
-  Koala_Finalize();
+  Vector vec;
+  Vector_Init(&vec);
 
-  fclose(in);
+  PackageInfo *pkg = New_PackageInfo(output);
+  ParserState *ps;
+  int errnum = 0;
+
+  struct dirent *dent;
+  char name[512];
+  FILE *in;
+  while ((dent = readdir(dir))) {
+    sprintf(name, "%s/%s", input, dent->d_name);
+    if (lstat(name, &sb) != 0) continue;
+    if (S_ISDIR(sb.st_mode)) continue;
+    in = fopen(name, "r");
+    if (!in) {
+      printf("%s: no such file or directory\n", name);
+      continue;
+    }
+    ps = new_parser(pkg, dent->d_name);
+    parser_enter_scope(ps, ps->sym->ptr, SCOPE_MODULE);
+    ps->u->sym = ps->sym;
+    yyset_in(in, ps->scanner);
+    yyparse(ps, ps->scanner);
+    fclose(in);
+    if (ps->errnum <= 0) {
+      Vector_Append(&vec, ps);
+    } else {
+      errnum += ps->errnum;
+      destroy_parser(ps);
+    }
+  }
+  closedir(dir);
+
+  Vector_ForEach(ps, &vec) {
+    parser_body(ps, &ps->stmts);
+    check_unused_imports(ps);
+    parser_exit_scope(ps);
+    //destroy_parser(ps);
+  }
+
+  if (errnum <= 0) {
+    codegen_klc(pkg);
+  } else {
+    fprintf(stderr, "There are %d errors.\n", errnum);
+  }
+
+  Vector_Fini(&vec, NULL, NULL);
+  Koala_Finalize();
 
   return 0;
 }
@@ -166,7 +218,7 @@ static int cmd_run(char *argv[], int size)
   puts(KOALA_START);
 
   Koala_Initialize();
-  Koala_Run(argv[0], "Main");
+  Koala_Run(argv[0], "main");
   Koala_Finalize();
 
   puts(KOALA_END);
